api: Rename 'subproject_items_reordered' - files
Rename the files and imports of 'subproject_items_reordered' to
'workflowitems_reordered'
