+++ api/src/service/domain/business_event.ts
@@ -21,7 +21,7 @@ import * as ProjectUpdated from "./workflow/project_updated";
 import * as SubprojectAssigned from "./workflow/subproject_assigned";
 import * as SubprojectClosed from "./workflow/subproject_closed";
 import * as SubprojectCreated from "./workflow/subproject_created";
-import * as SubprojectItemsReordered from "./workflow/subproject_items_reordered";
+import * as WorkflowitemsReordered from "./workflow/workflowitems_reordered";
 import * as SubprojectPermissionGranted from "./workflow/subproject_permission_granted";
 import * as SubprojectPermissionRevoked from "./workflow/subproject_permission_revoked";
 import * as SubprojectProjectedBudgetDeleted from "./workflow/subproject_projected_budget_deleted";
@@ -55,7 +55,7 @@ export type BusinessEvent =
   | SubprojectAssigned.Event
   | SubprojectClosed.Event
   | SubprojectCreated.Event
-  | SubprojectItemsReordered.Event
+  | WorkflowitemsReordered.Event
   | SubprojectPermissionGranted.Event
   | SubprojectPermissionRevoked.Event
   | SubprojectProjectedBudgetDeleted.Event
