+++ api/src/service/domain/workflow/subproject_eventsourcing.ts
@@ -8,7 +8,7 @@ import * as Subproject from "./subproject";
 import * as SubprojectAssigned from "./subproject_assigned";
 import * as SubprojectClosed from "./subproject_closed";
 import * as SubprojectCreated from "./subproject_created";
-import * as SubprojectItemsReordered from "./subproject_items_reordered";
+import * as WorkflowitemsReordered from "./workflowitems_reordered";
 import * as SubprojectPermissionGranted from "./subproject_permission_granted";
 import * as SubprojectPermissionRevoked from "./subproject_permission_revoked";
 import * as SubprojectProjectedBudgetDeleted from "./subproject_projected_budget_deleted";
@@ -63,7 +63,7 @@ function applySubprojectEvent(
       return apply(ctx, event, subprojects, event.subprojectId, SubprojectClosed);
 
     case "workflowitems_reordered":
-      return apply(ctx, event, subprojects, event.subprojectId, SubprojectItemsReordered);
+      return apply(ctx, event, subprojects, event.subprojectId, WorkflowitemsReordered);
 
     case "subproject_permission_granted":
       return apply(ctx, event, subprojects, event.subprojectId, SubprojectPermissionGranted);
